funcions splited from main file, function added to download received documments from telegram
